moving and reducing coordinate change
